requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
PyMuPDF>=1.25
python-docx>=1.1
python-multipart>=0.0.18
pydantic-settings>=2.7
chardet>=5.2
pytesseract>=0.3.13
Pillow>=11.0
